Create Program to display the Fibonacci sequence up to n-th term
A Fibonacci sequence is the integer sequence of 0, 1, 1, 2, 3, 5, 8....

The first two terms are 0 and 1. All other terms are obtained by adding the preceding two terms. This means to say the nth term is the sum of (n-1)th and (n-2)th term.

Here, we store the number of terms in num. We initialize the first term to 0 and the second term to 1.

If the number of terms is more than 2, we use a for loop to find the next term in the sequence by adding the preceding two terms. We then interchange the variables (update it) and continue on with the process.