Create Python Program to Find the Factors of a Number
To find the factors of another number, change the value of num.

In this program, the number whose factor is to be found is stored in num, which is passed to the factors() function. This value is assigned to the variable x in factors().

In the function, we use the for loop to iterate from i equal to x. If x is perfectly divisible by i, it's a factor of x.